Align help text with core django help text messages
This PR aligns the help text with the Django help text messages, which read like:

```
This password is too short. It must contain at least 12 characters.
This password is too common.
This password is entirely numeric.
```
This make this module more easily integrated with the core Django password validation rules.